+++ previous-versions/sync-for-commerce-version-1/src/codatsynccommerce/accounting_company_info.py
@@ -89,8 +89,8 @@ class AccountingCompanyInfo:
         
         if http_res.status_code == 200:
             if utils.match_content_type(content_type, 'application/json'):
-                out = utils.unmarshal_json(http_res.text, Optional[shared.AccountingDataset])
-                res.accounting_dataset = out
+                out = utils.unmarshal_json(http_res.text, Optional[shared.Dataset])
+                res.dataset = out
             else:
                 raise errors.SDKError(f'unknown content-type received: {content_type}', http_res.status_code, http_res.text, http_res)
         elif http_res.status_code in [401, 404, 429]: