+++ previous-versions/sync-for-commerce-version-1/src/codatsynccommerce/companies.py
@@ -107,7 +107,7 @@ class Companies:
         
         url = utils.generate_url(operations.UpdateCompanyRequest, base_url, '/companies/{companyId}', request)
         headers = {}
-        req_content_type, data, form = utils.serialize_request_body(request, "company_request_body", 'json')
+        req_content_type, data, form = utils.serialize_request_body(request, "create_company", 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
         headers['Accept'] = 'application/json'