+++ previous-versions/sync-for-commerce-version-1/src/codatsynccommerce/refresh_data.py
@@ -228,8 +228,8 @@ class RefreshData:
         
         if http_res.status_code == 200:
             if utils.match_content_type(content_type, 'application/json'):
-                out = utils.unmarshal_json(http_res.text, Optional[shared.DataConnectionHistory])
-                res.data_connection_history = out
+                out = utils.unmarshal_json(http_res.text, Optional[shared.PullOperations])
+                res.pull_operations = out
             else:
                 raise errors.SDKError(f'unknown content-type received: {content_type}', http_res.status_code, http_res.text, http_res)
         elif http_res.status_code in [400, 401, 404, 429]: